requirements.txt
absl-py==2.1.0
accelerate==1.4.0
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
audioread==3.0.1
babel==2.17.0
blis==1.2.0
blobfile==3.0.0
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
chex==0.1.89
click==8.1.8
cloudpathlib==0.21.0
colorama==0.4.6
confection==0.1.5
csvw==3.5.1
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cymem==2.0.11
datasets==3.3.2
decorator==5.2.1
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dlinfo==2.0.0
docopt==0.6.2
dotenv==0.9.9
espeakng-loader==0.2.4
etils==1.12.0
filelock==3.17.0
flax==0.10.4
frozenlist==1.5.0
fsspec==2024.12.0
greenlet==3.1.1
groq==0.18.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.28.1
humanize==4.12.1
idna==3.10
importlib_resources==6.5.2
isodate==0.7.2
jax==0.5.2
jaxlib==0.5.1
Jinja2==3.1.5
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kokoro==0.8.4
langchain==0.3.20
langchain-core==0.3.41
langchain-groq==0.2.4
langchain-text-splitters==0.3.6
langcodes==3.5.0
langsmith==0.3.11
language-tags==1.2.0
language_data==1.3.0
lazy_loader==0.4
librosa==0.10.2.post1
llama_cpp_python==0.3.7
llvmlite==0.44.0
loguru==0.7.3
lxml==5.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
misaki==0.8.4
ml_dtypes==0.5.1
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.12
nest-asyncio==1.6.0
networkx==3.4.2
num2words==0.5.14
numba==0.61.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opt_einsum==3.4.0
optax==0.2.4
orbax-checkpoint==0.11.6
orjson==3.10.15
packaging==24.2
pandas==2.2.3
phonemizer-fork==3.3.2
pillow==11.0.0
platformdirs==4.3.6
pooch==1.8.2
preshed==3.0.9
propcache==0.3.0
protobuf==6.30.0
psutil==6.1.1
pyarrow==19.0.1
pycparser==2.22
pycryptodomex==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyttsx3==2.98
pytz==2025.1
PyYAML==6.0.2
rdflib==7.1.3
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==1.5.0
rich==13.9.4
rpds-py==0.23.1
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.2
segments==2.3.0
sentencepiece==0.2.0
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.13.1
soxr==0.5.0.post1
spacy==3.8.4
spacy-curated-transformers==0.3.0
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.38
srsly==2.5.1
sympy==1.13.1
tenacity==9.0.0
tensorstore==0.1.72
thinc==8.3.4
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.0
toolz==1.0.0
torch==2.6.0+cpu
torchaudio==2.6.0+cpu
torchvision==0.21.0+cpu
tqdm==4.67.1
transformers==4.49.0
treescope==0.1.9
triton==3.2.0
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
